Added option for voice only installation for higurashi. The dropbox in the installation menu is now configurable only for higurashi, Umineko isntallation grid stays the same.
A bat file for cleaning the temp folder has been added to the resources zip, so now, after the user clicks the button "Finish", that bat will be called. This is only for higurashi, Umineko stays the same.

Created a new variable in the patcherPOCO: "Patcher.BatName" the program sets it when the patcher object is being constructed in the confimation menu.